Add dirty cache reuse

+++ deliverance-core/src/main/java/io/teknek/deliverance/tensor/FloatBufferTensor.java
@@ -104,8 +104,8 @@ public final class FloatBufferTensor extends AbstractTensor<FloatVector, Float> 
     @Override
     public void copyFrom(AbstractTensor src, int srcOffset, int destOffset, int length) {
         Preconditions.checkArgument(this.dType == src.dType, "Different types");
-        segment.asSlice(getMemorySegmentOffset(destOffset),  length * dType.size())
-                .copyFrom(src.getMemorySegment().asSlice(src.getMemorySegmentOffset(srcOffset),  length * dType.size()));
+        segment.asSlice(getMemorySegmentOffset(destOffset), (long) length * dType.size())
+                .copyFrom(src.getMemorySegment().asSlice(src.getMemorySegmentOffset(srcOffset), (long) length * dType.size()));
     }
 
     @Override
